requirements.txt
PyQt6>=6.0.0
Pillow>=8.0.0
reportlab>=3.5.0
